Just some code cleanup

+++ include/nn/atk/atk_SequenceSoundFile.h
@@ -27,7 +27,7 @@ struct SequenceSoundFile {
     struct LabelBlockBody {
         Util::ReferenceTable labelInfoReferenceTable;
 
-        int GetLabelCount() const { return labelInfoReferenceTable.count; }
+        int GetLabelCount() const { return static_cast<int>(labelInfoReferenceTable.count); }
 
         const LabelInfo* GetLabelInfo(int index) const;
         const char* GetLabel(int index) const;
